fixed integer types

+++ src/new/astar.h
@@ -5,8 +5,6 @@
 #ifndef WARZONE2100_ASTAR_H
 #define WARZONE2100_ASTAR_H
 
-#include <cstdint>
-
 enum class ASTAR_RETVAL
 {
   OK,
@@ -18,19 +16,19 @@ class Path_Coordinate
 {
 public:
   Path_Coordinate() = default;
-  Path_Coordinate(int16_t x, int16_t y);
+  Path_Coordinate(int x, int y);
 
   bool operator ==(const Path_Coordinate &rhs) const;
   bool operator !=(const Path_Coordinate &rhs) const;
 private:
-  int16_t x, y;
+  int x, y;
 };
 
 struct Path_Node
 {
   Path_Coordinate path_coordinate;
-  uint32_t distance_from_start;
-  uint32_t estimated_distance_to_end;
+  unsigned distance_from_start;
+  unsigned estimated_distance_to_end;
 };
 
 #endif // WARZONE2100_ASTAR_H
